requirements.txt
django==3.2.0
djangorestframework==3.12.0
djangorestframework-simplejwt==5.1.0